applied patch from Danilo Piazzalunga
mail text:
Trying to compile uwadv with gcc 4.0 fails because new versions
of gcc do not support anymore the cast-as-lvalue extension (see
also http://gcc.gnu.org/gcc-3.4/changes.html).

patch comment:
IIRC, casting an lvalue is a dire error in C (it is allowed as a gcc
extension up to gcc 3.3). A cast from any pointer to a void* is always
safe. The new macro may not have a pretty name but it does the same thing
while preserving the encapsulation of the SDL_rwops internals.